move api package to inside node package
