+++ node/api/host_test.go
