+++ node/api/miner_test.go
