+++ node/api/scan.go
